gcc 3.4.x picky at type-casting from int to pointer or reverse, we should bridge the type casting with INTPTR -- the integer with the same size of pointer
git-svn-id: http://svn.open64.net/svnroot/open64@317 ead328b1-e903-0410-a19d-eacee4b031f6

+++ projects/pathscale-2.2.1/kpro64/common/util/resource.c
@@ -203,7 +203,7 @@ Get_Resources (
 #endif
 
     /* Get the memory information */
-    r->memory = (INT) sbrk(0);
+    r->memory = (INT64)sbrk(0);
     r->freemem = 0;
 }
