+++ projects/pathscale-2.2.1/kpro64/libcmplrs/make_depend.c
@@ -693,7 +693,7 @@ MDclose(MDhandle h, char *target)
 	if ( 0 < md.stb.st_size ) {
 		md.base = mmap	( 0, md.stb.st_size, PROT_READ|PROT_WRITE
 				, MAP_SHARED, md.f, 0);
-		if ((int) md.base < 0) 
+		if ((int)(long)md.base < 0) 
 			ERR(("mmap (%s): %s", md.filename, strerror(errno)));
 		md.size  = md.stb.st_size;
 		md.limit = md.base + md.size;
